Add metrics for number of inodes (#781)
* Make metrics gauges remember if they've changed

This is a small change to their behavior: they'll now emit if they're
changed to zero, whereas before they would not. But I think that's the
behavior we want anyway.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Convert absolute counters to gauges

We were abusing absolute counters to get the "only emit when changed"
behavior, which gauges now provide too.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Add metrics for number of inodes

We track the total number of inodes as well as tracking them by kind.

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>