Don't use AutoUnmount in FUSE integration tests
It's the wrong thing to do: the FUSE session is unmounted automatically
when the BackgroundSession drops. AutoUnmount is for the case when the
mounting process might not unmount on its own, but we always do. Using
the option spawns a new thread and leaks a socket until the end of the
process, which means our test processes can run out of open file
descriptors if they run many FUSE sessions.

Signed-off-by: James Bornholt <bornholt@amazon.com>