Update to new `metrics` crate version (#787)
v0.22.0 of the metrics crate was a breaking change to how its macros
work -- they now return the counter/gauge/histogram itself and you call
methods on it to record metrics, rather than recording the metric as
part of the macro. So this change is mostly a find and replace to get
things compiling again with this new change.

Other than that change, there's two new things we'd like to use from
this new version:

1. Scoped local metrics recorders are now supported, which makes it much
   easier to write unit tests for metrics. The metrics recorders were
   previously global, so tests had to use `rusty_fork` to fork a new
   test process. I've used this change to update the current basic
   metrics tests.
2. Metrics now include metadata such as severity and module/line
   location. We're not using this yet, but could use it in the future to
   create scoped metrics or different metric severities for our logging
   use.

Signed-off-by: James Bornholt <bornholt@amazon.com>