Cancel S3 requests when dropped
Today we don't cancel S3 requests when dropped. For our prefetcher that
means we keep streaming (up to) 2GB of data that will never be used.
This change cancels in-flight requests when dropped, so that the CRT
will stop streaming them. Some bytes might still be in flight or
delivered, which is fine. Canceling requests is a no-op if they've
already completed.

The tricky case for this change is PutObject. Our current implementation
of `PutObjectRequest::write` blocks until the bytes it provides are
consumed by the client. But sometimes the client might stop reading from
the stream because the request has failed. That case happens to work
today because we don't retain a reference to the meta request ourselves,
and so the failed request's destructors run immediately after the
failure, which unblocks the writer and returns it an error. But now we do
hold onto a reference, and the destructors can't run until the last
reference is released, so the writer is never unblocked. To fix this, we
make the `write` and `complete` methods of the `PutObjectRequest` poll
_both_ the write stream and the request itself in parallel. If the request
completes, this gives us a chance to bail out of the write rather than
blocking forever.

Signed-off-by: James Bornholt <bornholt@amazon.com>