Remove crate check workflow for non -sys crates
This workflow exists because we're worried about the 10MiB limit. The
sys crate is the only one that's in any danger of hitting that limit
because it's packaging a ton of C code, including an entire TLS
implementation. This workflow is also broken (#700) when updating crate
versions for multiple crates at the same time, which we really want to
do.

Since we only care about the size of the sys crate, let's only check
that one.

Signed-off-by: James Bornholt <bornholt@amazon.com>