First pass at randomized testing for POSIX semantics
The idea is to randomly run real system calls against both a real file
system (a temporary directory on the OS disk) and Mountpoint. We expect
Mountpoint to allow fewer behaviors than a real POSIX file system (e.g.
can't read and write the same file handle), so we need some logic for
allowing a limited set of divergences between the two systems.

Signed-off-by: James Bornholt <bornholt@amazon.com>