First pass at randomized testing for POSIX semantics (#842)
* Don't use AutoUnmount in FUSE integration tests

It's the wrong thing to do: the FUSE session is unmounted automatically
when the BackgroundSession drops. AutoUnmount is for the case when the
mounting process might not unmount on its own, but we always do. Using
the option spawns a new thread and leaks a socket until the end of the
process, which means our test processes can run out of open file
descriptors if they run many FUSE sessions.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* First pass at randomized testing for POSIX semantics

The idea is to randomly run real system calls against both a real file
system (a temporary directory on the OS disk) and Mountpoint. We expect
Mountpoint to allow fewer behaviors than a real POSIX file system (e.g.
can't read and write the same file handle), so we need some logic for
allowing a limited set of divergences between the two systems.

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>