Small refactors of inode code (#888)
* Remove process ID from Inode

The entire check happens in S3Filesystem anyway, so there's no need for
the actual Inode to be aware of the PID.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Normalize WriteHandle/ReadHandle in inode.rs

There's no reason for start_writing to exist since it's always called
immediately on the result of write. Once that change is made, it looks
helpful to have a symmetric ReadHandle for the lifecycle of reads, even
though it doesn't do much today.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Move more stuff off the Inode

`inc_file_size` belongs on the write handle itself (you have to be
writing to increment the file size), and some other stuff here doesn't
need to be public.

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>