Reset prefetcher on any error (#933)
* Add test of PartQueue invariant failure

Signed-off-by: Alessandro Passaro <alexpax@amazon.com>

* Reset prefetcher on any error

The prefetcher did not reset its internal state (and cancel current tasks) when encountering an error on a forward seek.
As a result, successive reads could try and read from a part queue in an invalid state. This change ensures that the prefetcher
is always reset when encountering an error, whether while reading or seeking.

Signed-off-by: Alessandro Passaro <alexpax@amazon.com>

* Fix prefetcher reset on integrity error

Signed-off-by: Alessandro Passaro <alexpax@amazon.com>

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.com>
Co-authored-by: Alessandro Passaro <alexpax@amazon.com>