Re-implement the prefetcher using backpressure mechanism (#980)
* Re-implement the prefetcher using backpressure mechanism

The prefetcher now uses only one GetObject request to fetch data in advance.
This request has a range of entire object but use backpressure mechanism
to control how much data it wants to fetch into the part queue instead of
spawning up to two requests in parallel.

This should make the throughput more stable because previously the two
request tasks could compete with each other when fetching data from S3.
Also, it will be easier to control how much data we want to store in the
part queue.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* PR comments

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Fix an issue where EmptyReadWindow error could be reported when request is already completed

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* PR comments

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

---------

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>