Avoid extending part on backward seek (#1005)
* Avoid extending part on backward seek

Currently, we combine parts from the seek window to `current_part` in the
part queue whenever we seek backward which mean we also have to re-compute
checksums for this combined part. It particularly affect read throughput in
some use cases where backward seek rate is high. This change should improve
the throughput for those use cases.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* PR comments

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

---------

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>