Update non-existent network interface initialization test
CRT was returning error during first operation before if it provided
with a non-existent network interface name. But with
https://github.com/awslabs/aws-c-s3/pull/456, it started failing
during the client creation phase. Our tests were written for the
previous behaviour and was expecting client creation to succeed even
with an invalid network interface. The test is updated to expect
errors during client creation.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>