Update PR template to reflect how PR metadata is used in PR merges (#1090)
## Description of change

We have updated pull request settings to use the description of the PR
in the commit message rather than the list of individual commits in the
pull request.

This change explains the importance of the title and description in how
they are used to create the squash commit. We want to encourage PR
authors to be aware of this and keep the metadata up-to-date.

Relevant issues: N/A

## Does this change impact existing behavior?

No, PR template change only.

## Does this change need a changelog entry in any of the crates?

No, there are no code changes.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>