Add checksum algorithm to ListObjectsV2 response (#1086)
## Description of change

For a consumer of the S3 client, we need to return the checksum
algorithm used with objects.

This change exposes the checksum algorithm that is sometimes returned as
part of a ListObjectsV2 request. This functionality is not opt-in, and
will now simply be exposed to be used by consumers of Mountpoint S3
client where supported.

This change also updates `ChecksumAlgorithm` to have a new `Unknown`
variant. This is a common pattern in AWS SDKs. This will allow clients
to recognize where an unknown algorithm is returned (should the S3
service start supporting additional algorithms), and avoid either
returning `None` (if optional) or panicking when the error may be
recoverable.

Relevant issues: N/A

## Does this change impact existing behavior?

Yes, it changes the `ObjectInfo` and `ChecksumAlgorithm` structs.
Specifically, they are now marked `non_exhaustive` meaning that new
fields could be added in the future (as has been done in this PR).

## Does this change need a changelog entry in any of the crates?

Yes, relevant change logs for breaking change (`non_exhaustive`) as well
as new feature (exposing checksum algorithm) have been added to
`mountpoint-s3-client`'s changelog.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>