Update ChecksumAlgorithm field of client's ObjectInfo struct to be Vec over single element (#1093)
## Description of change

On reviewing the S3 API documentation, the checksum algorithm field is a
list of algorithms. Additionally, when reviewing other SDKs such as the
[Rust SDK, we see that they are presenting this field as
`Option<Vec<String>>`](https://docs.rs/aws-sdk-s3/latest/aws_sdk_s3/types/struct.Object.html)
rather than a single optional element. (Note, we do drop the `Option`
still.

We'd prefer to align with the SDK interface. Our tenet here is to ensure
our S3 client is consistent with the official SDKs where there's no
significant effort required. This is making a breaking change while
we're already planning to make a number of breaking changes to the
client.

Relevant issues:

- Follow up on #1086, which added checksum algorithms to the list
objects response.

## Does this change impact existing behavior?

Yes, it changes the S3 client behavior to return a different type. We
are however merging this before a new crate release, so this will not be
an additional breaking change.

There's no behavior change to Mountpoint file system.

## Does this change need a changelog entry in any of the crates?

There is already an existing entry in `mountpoint-s3-client`'s
changelog. This PR has been added to the list of PRs for that entry.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>