Update macOS unit tests to use latest macOS (#1097)
## Description of change

This change updates the CI to use the latest version of macOS available
on GitHub runners (currently macOS 15), updating from macOS 12.

We are seeing GitHub CI failing in some cases. We wanted to update
version anyway, and I suspect the issues we are seeing is due to the
deprecation (maybe).

I update to latest since we don't strictly support macOS, so just
checking that unit tests pass on the latest version is enough for us.

Relevant issues:
- https://github.com/github/roadmap/issues/986

## Does this change impact existing behavior?

CI change only. No behavior change.

## Does this change need a changelog entry in any of the crates?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>