Decouple S3FileSystem and S3FuseFileSystem constructors (#1100)
<!--
The title and description of pull requests will be used when creating a
squash commit to the base branch (usually `main`).
Please keep them both up-to-date as the code change evolves, to ensure
that the commit message is useful for future readers.
-->

## Description of change

Refactor `S3FuseFileSystem::new` to accept an `S3FileSystem` instance
rather than creating a new one. The change highlights that
`S3FuseFileSystem` is only a wrapper for `S3FileSystem` and will make it
easier to modify `S3FileSystem` construction in future changes.

## Does this change impact existing behavior?

No, it's only a minor internal refactor.

## Does this change need a changelog entry in any of the crates?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>