Implement checksums for MockClient upload path (#1102)
## Description of change

This change updates the mock client's write path to compute and store
checksums, matching behavior approximate to S3. We want this so that we
can use the mock client for both uploads and downloads and verify
checksum behavior for the client.

The change stores the checksums as strings, as this is the observed
behavior of S3. We can always update if this turns out to be a bad
assumption.

Relevant issues: N/A

## Does this change impact existing behavior?

This change updates the `mountpoint-s3-client` mock client to add
checksum persistence to the write path. This was previously a gap in the
mock client.

## Does this change need a changelog entry in any of the crates?

I do not think this needs a changelog entry, as it does not change the
behavior of the S3 client itself. Happy to discuss.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>