Lay out files using fio in the benchmark.
This makes the benchmarks more self-contained, creating the state that
they need to run rather than relying on pre-created state (potentially
with different/unknown mount options and/or object properties).

It does change behaviour a little: previously the multi-thread tests
would use the same object whereas now each thread uses its own object --
arguably this is a more useful test but results in higher S3 usage.

Also note that the cache benchmark is unchanged in this commit, because
it makes assumptions about the filename used by the tests.

Removing the assumption that each test will operate on a single file
prepares us for future mixed read/write tests, and allows different fio
jobs to be run in parallel safely.

Signed-off-by: Andrew Peace <adpeace@amazon.com>