Don't run benchmarks with debug, plus minor refactoring (#1104)
## Description of change

Remove the `--debug` flag when mounting S3 for the benchmarks.

This is now available via an S3_DEBUG environment variable which, when
set, will add `--debug` back again.

Also, within `fs_bench.sh`, merge the read and write benchmark methods
into a single one, which is paramterized, since they were almost
identical. This avoids having to make the change described above in two
places and simplifies the code going forwards.

## Does this change impact existing behavior?

This changes the benchmarks to run without --debug to the mount command,
which creates a discontinuity in benchmark results, and may improve them
(though there's no actual performance improvement here).

## Does this change need a changelog entry in any of the crates?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Andrew Peace <adpeace@amazon.com>