Add missing documentation for fuser Mount structs (#1105)
Our fuser fork emits a lot of warnings during builds, impacting our pull
requests. This change addresses missing documentation on one struct in
fuser (which was introduced by our change in
https://github.com/awslabs/mountpoint-s3/pull/1098).

This change is a good candidate to be contributed to upstream. Once
contributed, this commit can be removed.

Relevant issues: N/A

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>