Add descriptive name to benchmarks, enable CI summaries (#1109)
## Description of change

This change leverages our benchmark action's GitHub Action's job summary
integration to display a summary for throughput and latency benchmarks.
Before this change, we would need to check the webpage or the JSON
output to compare. For pull requests, only the JSON output would be
available. This change applies to both commits on `main` as well as pull
requests.

This should also address issues we had in the past where commit comments
did not describe which benchmark failed. See implementation:
https://github.com/benchmark-action/github-action-benchmark/blob/6bae118c112083251560ad8b3a1ff2e43aa23351/src/write.ts#L203-L239

More information on benchmark GH Action's integration:
https://github.com/benchmark-action/github-action-benchmark?tab=readme-ov-file#job-summary

Relevant issues: N/A

## Does this change impact existing behavior?

No impact to Mountpoint file system or client.

## Does this change need a changelog entry in any of the crates?

No, no customer-facing change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>