Add GitHub runner architecture to CI cache key (#1110)
<!--
The title and description of pull requests will be used when creating a
squash commit to the base branch (usually `main`).
Please keep them both up-to-date as the code change evolves, to ensure
that the commit message is useful for future readers.
-->

## Description of change

This change adds the architecture of the runner to the cache key. In a
previous change where we upgraded macOS runners from macOS 12 to macOS
15, the architecture changed however the cache was not invalidated.

Likely we should find a way to key the cache on the actual operating
system version used however there doesn't seem to be an obvious way to
do that right now. This quick fix should at least protect us from
architecture changes (which macOS runners do have, but I understand its
only for larger runners at this time).

Relevant issues: #1097

## Does this change impact existing behavior?

No, only CI change.

## Does this change need a changelog entry in any of the crates?

No, only CI change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>