Update PyTorch example with clearer getting started section (#1092)
<!--
The title and description of pull requests will be used when creating a
squash commit to the base branch (usually `main`).
Please keep them both up-to-date as the code change evolves, to ensure
that the commit message is useful for future readers.
-->

## Description of change

I've been using this example for some talks I've been preparing. I'm
updating the steps to be clearer on what action needs to be taken by the
reader (since I kept missing steps when preparing new instances).

In this change, I also update the mentioned bucket names to follow the
example bucket names recommended across AWS documentation.

Relevant issues: N/A

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>