Ensure that the file system in tests is unmounted before removing the temp dir (#1116)
## Description of change

Ensure that the file system in integration tests is unmounted before
removing the temporary directory. The introduction of `TestSession` in
#1096 inadvertently changed the order in which the temporary directory
and the FUSE session are dropped. Previously it was hidden in the
declaration order. This change makes it explicit in `drop`.

## Does this change impact existing behavior?

No.

## Does this change need a changelog entry in any of the crates?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>