Enable metric emission in benchmark runs (#1120)
## Description of change

We opted to disable debug logging in #1104 as this may impact
performance, however it was not known that the memory usage monitoring
was dependent on metrics being emitted implicitly due to `--debug`. This
change restores metrics in logs to fix the peak memory usage
benchmarking.

Relevant issues: #1104

## Does this change impact existing behavior?

No change to Mountpoint.

## Does this change need a changelog entry in any of the crates?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>