Add support for getting object checksums in GetObject (#1123)
<!--
The title and description of pull requests will be used when creating a
squash commit to the base branch (usually `main`).
Please keep them both up-to-date as the code change evolves, to ensure
that the commit message is useful for future readers.
-->

## Description of change

- Adds new method `get_object_checksum` to `get_object` to retrieve the
whole object checksum.
- Adds new parameter `checksum_mode` to `GetObjectParams` to configure
if we want to request checksums from S3.
- If checksums aren't requested, `get_object_checksum` returns an error.
- Refactor `S3CrtClient` to store a cell of the object headers rather
than metadata.

<!--
    Please describe your contribution here.
    What is the change and why are you making it?
-->

Relevant issues: <!-- Please add issue numbers. -->

## Does this change impact existing behavior?

No

<!-- Please confirm there's no breaking change, or call our any behavior
changes you think are necessary. -->

## Does this change need a changelog entry in any of the crates?

Yes

<!--
    Please confirm yes or no.
    If no, add justification. If unsure, ask a reviewer.

    You can find the changelog for each crate here:
-
https://github.com/awslabs/mountpoint-s3/blob/main/mountpoint-s3/CHANGELOG.md
-
https://github.com/awslabs/mountpoint-s3/blob/main/mountpoint-s3-client/CHANGELOG.md
-
https://github.com/awslabs/mountpoint-s3/blob/main/mountpoint-s3-crt/CHANGELOG.md
-
https://github.com/awslabs/mountpoint-s3/blob/main/mountpoint-s3-crt-sys/CHANGELOG.md
-->

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Simon Beal <simobeal@amazon.com>