Update vendored fuser to daad5673 (#1129)
## Description of change

This change updates the vendored fuser version. The changes include an
amended README to include information on how to maintain the fork, as
well as rebasing our patches on top of the upstream fuser repository.

The main changes we're interested in here is to eliminate many of the
build warnings that are currently showing up in pull requests.

Relevant issues: N/A

## Does this change impact existing behavior?

No change in behavior of any crate.

## Does this change need a changelog entry in any of the crates?

No change log needed.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>