Bypass the shared cache for large objects (#1117)
## Description of change

This change makes `get_block` and `put_block` for objects larger than
`1MiB` be a no-op in the shared cache.

Relevant issues: N/A

## Does this change impact existing behavior?

No, it is under the feature flag.

## Does this change need a changelog entry in any of the crates?

Yes, in the following PRs.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>