Add mixed read/write tests to the benchmark. (#1130)
## Description of change

As we make changes that might impact mixed read/write workloads it is
useful to have some examples of these in our benchmarks. This change
adds a 20/80, 50/50, and 80/20 read/write workload to the benchmarks.

The results reporting is updated to support multiple job types in a
single benchmark. This works by continuing to average over iterations as
before, but averages each job separately then sums the averages to
produce the final throughput number for that benchmark (i.e. (avg(read
throughput) + avg(write throughput)) for the mixed benchmarks).

## Does this change impact existing behavior?

No.

## Does this change need a changelog entry in any of the crates?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Andrew Peace <adpeace@amazon.com>