Verify object metadata in Express cache (#1125)
## Description of change

- Verify S3 Express cache objects have valid object metadata which
matches the keys
- Verifies the CRC32C of the object content post-download from S3
  - If checksum is missing, return `BlockChecksumMissing`.

Relevant issues: N/A

## Does this change impact existing behavior?

Changes S3 Express cache to require object metadata and CRC32C to be
present. Old caches will not be used.

## Does this change need a changelog entry in any of the crates?

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Simon Beal <simobeal@amazon.com>