Rename the shared cache CLI flag (#1144)
## Description of change

Rename the CLI flag for the shared cache. New help message:

```bash
--cache-xz <BUCKET>
    Enable caching of object content to the specified bucket on S3 Express One Zone (same region only)
```

Relevant issues: N/A

## Does this change impact existing behavior?

No, it's behind a feature flag.

## Does this change need a changelog entry in any of the crates?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>